docs(components): [transfer] improve and complete examples
